Update Trip and Itinerary data models and synchronization logic
- Change `userId` type from `Int` to `String` in `Trip` model, `TripRepository`, and `TripDao` to support Firebase UIDs.
- Implement foreign key constraints with `CASCADE` delete for `ItineraryItem` (linked to `Trip`) and `Trip` (linked to `User`).
- Add `getLoggedUID()` to `AuthRepository` to retrieve the current Firebase user ID.
- Update `TripViewModel` to automatically fetch the authenticated `userId` when creating a new trip.
- Remove manual itinerary item deletion in `TripRepositoryImpl`, relying on database-level cascade deletes instead.
- Remove `deleteItemsByTrip` from `ItineraryDao`.
- Fix a type casting bug in `ItineraryViewModel` when processing repository results.

+++ app/app/src/main/java/com/monarca/smarttravel/ui/viewmodels/ItineraryViewModel.kt
@@ -94,7 +94,7 @@ class ItineraryViewModel @Inject constructor(
             try {
                 val result = repository.addItineraryItem(newItem)
                 Log.i(TAG, "addItem: item creat correctament")
-                status = AppError.fromCode(result.toInt())
+                status = AppError.fromCode(result)
             } catch (e: Exception) {
                 Log.e(TAG, "addItem: error al crear item", e)
                 status = AppError.UNKNOWN
